Fixed int overflow in perft tests
i'm way too embarased to say how long it took to figure out that was the
problem...

+++ include/engine.hpp
@@ -23,7 +23,7 @@ public:
 
   bool make_move(const Move &move);
   void undo_move();
-  int perft(int depth);
+  uint64_t perft(int depth);
   std::vector<std::pair<Move, int>> perft_divide(int depth);
 
 private:
